+++ gldb/query/rdfstorequery.py
@@ -2,7 +2,7 @@ from abc import ABC
 
 import rdflib
 
-from gldb.query.query import Query
+from gldb.query.query import Query, QueryResult
 
 
 class RDFStoreQuery(Query, ABC):
@@ -11,21 +11,14 @@ class RDFStoreQuery(Query, ABC):
 
 class SparqlQuery(RDFStoreQuery):
 
-    def __init__(self, sparql_query: str, description: str="", *args, **kwargs):
-        self._args = args
-        self._kwargs = kwargs
-        self._description = description
-        self.sparql_query = sparql_query
-
-    @property
-    def description(self):
-        return self._description
-
     def __repr__(self):
-        return f"{self.__class__.__name__}({self.sparql_query!r})"
+        return f"{self.__class__.__name__}({self.query!r})"
 
     def __str__(self):
-        return self.sparql_query
+        return self.query
 
-    def execute(self, graph: rdflib.Graph) -> rdflib.query.Result:
-        return graph.query(self.sparql_query, *self._args, **self._kwargs)
+    def execute(self, graph: rdflib.Graph) -> QueryResult:
+        return QueryResult(
+            query=self,
+            data=graph.query(self.query, *self._args, **self._kwargs)
+        )