+++ tests/test_query.py
@@ -6,21 +6,15 @@ from gldb.query import Query, QueryResult
 class TestVersion(unittest.TestCase):
 
     def test_query(self):
-        class SQLQuery(Query):
-            def __init__(self, query: str, description: str):
-                self.query = query
-                self._description = description
 
-            @property
-            def description(self):
-                return self._description
+        class SQLQuery(Query):
 
             def execute(self, *args, **kwargs) -> QueryResult:
                 return QueryResult(self, "result")
 
         q = SQLQuery("SELECT * FROM Customers;", "Get all customers")
-        assert q.query == "SELECT * FROM Customers;"
-        assert q.description == "Get all customers"
+        self.assertEqual(q.query, "SELECT * FROM Customers;")
+        self.assertEqual(q.description, "Get all customers")
 
         res = q.execute()
-        assert isinstance(res, QueryResult)
+        self.assertIsInstance(res, QueryResult)